refactor: 패키지 구조 변경
- JwtTokenProvider, PasswordEncoder common으로 이동
- Image 패키지 common으로 이동
